% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/changeDecisionsDummies.R
\name{changeDecisionsDummies}
\alias{changeDecisionsDummies}
\title{Takes decisions and modifies them so that only one level of a multiclass variable is used in decisions}
\usage{
changeDecisionsDummies(
  rules,
  dummy_var,
  data,
  target,
  classPos = NULL,
  in_parallel = FALSE,
  n_cores = detectCores() - 1,
  cluster = NULL
)
}
\arguments{
\item{rules}{a data frame with a column "condition".}

\item{dummy_var}{string vector with the names of columns to change to dummy variable.}

\item{data}{data on which to fit the decision ensemble.}

\item{target}{response variable.}

\item{classPos}{for classification, the positive class.}

\item{in_parallel}{if TRUE, the function is run in parallel.}

\item{n_cores}{if in_parallel = TRUE, and no cluster has been passed: number of cores to use, default is detectCores() - 1.}

\item{cluster}{the cluster to use to run the function in parallel (opt).}
}
\description{
Takes decisions and modifies them so that only one level of a multiclass variable is used in decisions
}
