% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Misc.R
\name{compatibleNames}
\alias{compatibleNames}
\title{Transform character strings to be compatible with endoR functions.}
\usage{
compatibleNames(x)
}
\arguments{
\item{x}{character string or vector}
}
\description{
Transform character strings to be compatible with endoR functions.
}
\examples{
x <- c('hello.world__I am;happy', 'me-2')
compatibleNames(x)
}
