% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decisionImportance.R
\name{decisionImportance}
\alias{decisionImportance}
\title{Compute the importance of decisions}
\usage{
decisionImportance(
  rules,
  data,
  target,
  classPos = NULL,
  in_parallel = FALSE,
  n_cores = detectCores() - 1,
  cluster = NULL
)
}
\arguments{
\item{data}{data from which to get the decision support.}

\item{target}{response variable.}

\item{classPos}{if classification, the positive class.}

\item{in_parallel}{if TRUE, the function is run in parallel.}

\item{n_cores}{if in_parallel = TRUE, and no cluster has been passed: number of cores to use.}

\item{cluster}{the cluster to use to run the function in parallel.}

\item{rulesa}{a data.frame with a column "condition" or a vector with name "condition".}
}
\value{
the data.frame passed in rules with the gain and importance of the each decision.
}
\description{
This function computes the importance of decisions.
}
