% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discretizeData.R
\name{discretizeData}
\alias{discretizeData}
\title{Discretize numerical variables in a dataset}
\usage{
discretizeData(data, conditions, Kmax = 2, return_split = FALSE)
}
\arguments{
\item{data}{data to discretize.}

\item{Kmax}{numeric, maximal number of categories for each variable (default: K = 2).}

\item{return_split}{if TRUE, then the table with thresholds used to discretize data is also returned.}
}
\value{
Data with discretized variables.
}
\description{
This function discretizes all numerical variables, or only the ones passed in features, into K categories based on their quantiles. 
Names of categories can optionally be given in knames.
}
