% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discretizeData_model.R
\name{discretizeData_model}
\alias{discretizeData_model}
\title{Discretize numerical variables in a dataset based on thresholds used in the model to create splits.}
\usage{
discretizeData_model(data, conditions, Kmax = 2, return_split = FALSE)
}
\arguments{
\item{data}{data to discretize}

\item{conditions}{character vector with all conditions from which to find the thresholds}

\item{Kmax}{numeric, maximal number of categories for each variable (default: Kmax = 2).}

\item{return_split}{if TRUE, then the table with thresholds used to discretize data is also returned.}
}
\value{
Data with discretized variables.
}
\description{
This function discretizes all numerical variables into Kmax categories based on the splits used by the model for each variable.
}
