% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endoR.R
\docType{package}
\name{endoR}
\alias{endoR}
\title{endoR}
\description{
endoR extracts and visualizes how predictive variables contribute to tree ensemble model accuracy.
}
\author{
Albane Ruaud \email{albane.ruaud@tuebingen.mpg.de}
}
