% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featureImportance.R
\name{featureImportance}
\alias{featureImportance}
\title{Calculate the feature importances from a decision ensemble}
\usage{
featureImportance(nodes_agg)
}
\arguments{
\item{nodes_agg}{a datatable with, for variable in each decision, their name (column Feature), decision-wise importance (column importance), the importance and multiplicity of the decision (columns imp and n).}
}
\value{
a datatable with the feature importance for each variable.
}
\description{
This function calculates the featureImportance of variables in a decision ensemble. Levels of discretized variables are grouped so that only one importance is returned ('variable name' separated by 2*'_' from the 'level name').
}
