% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getThresholds.R
\name{getThresholds}
\alias{getThresholds}
\title{Get thresholds to discretize variables according to their split in tree ensemble models}
\usage{
getThresholds(conditions, data, Kmax = 2)
}
\arguments{
\item{conditions}{character vector with all conditions from which to find the thresholds}

\item{data}{data to discretize}

\item{Kmax}{numeric, maximal number of categories for each variable (default: Kmax = 2).}
}
\description{
Get thresholds to discretize variables according to their split in tree ensemble models
}
