% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFeatures.R
\name{plotFeatures}
\alias{plotFeatures}
\title{Plot the importance and influence of features.}
\usage{
plotFeatures(decision_ensemble, pi_thr, levels_order = NULL)
}
\arguments{
\item{decision_ensemble}{stable decision ensemble (see stabilitySelection).}

\item{pi_thr}{pi threshold to subset decisions according to their probability of being selected}

\item{levels_order}{optional, order for variables levels on the influence plot}
}
\value{
2 ggplots arranged in a row with ggpubr.
}
\description{
Returns a ggplot object with variables importance (across all categorical levels for factor variables) and variable per-level influence. 
It uses the ggpubr package to combine plots.
}
