% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotNetwork.R
\name{plotNetwork}
\alias{plotNetwork}
\title{Plot an interaction variable network.}
\usage{
plotNetwork(
  decision_ensemble,
  n_edges = NULL,
  layout = "stress",
  colour_edge_low = "firebrick2",
  colour_edge_mid = "grey90",
  colour_edge_high = "dodgerblue3",
  colour_node_low = "firebrick2",
  colour_node_mid = "grey90",
  colour_node_high = "dodgerblue3",
  text_size = 4,
  hide_nodes = 0,
  seed = 0
)
}
\arguments{
\item{decision_ensemble}{stable decision ensemble: list with edges, nodes, etc.}

\item{n_edges}{number of edges to plot; if NULL, all edges are plotted (default).}

\item{layout}{network layout, default is 'stress' (see ggraph package)}

\item{text_size}{size of node labels.}

\item{hide_nodes}{numeric, the minimal connectivity of a node. If negative, all nodes are plotted.}

\item{seed}{the seed to use for generating the network.}

\item{colour_x_y}{colour of x = edge or node, and y = low, mid or high (e.g., colour_edge_mid), to use for the colour gradients.}
}
\value{
a ggraph object
}
\description{
Returns a plot of nodes and edges. Plots are created with the ggraph and igraph packages.
}
