% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model2DE_resampling.R
\name{model2DE_resampling}
\alias{model2DE_resampling}
\title{Run model2DE on several bootstrap resamples.}
\usage{
model2DE_resampling(
  model,
  model_type,
  data,
  target,
  classPos = NULL,
  times = 10,
  p = 0.5,
  sample_weight = sample_weight,
  ntree = "all",
  maxdepth = Inf,
  dummy_var = NULL,
  prune = TRUE,
  maxDecay = 0.05,
  typeDecay = 2,
  discretize = TRUE,
  K = 2,
  features_ctg = NULL,
  filter = TRUE,
  min_imp = 0.9,
  seed = 0,
  in_parallel = FALSE,
  n_cores = detectCores() - 1,
  cluster = NULL
)
}
\arguments{
\item{times}{number of bootstraps}

\item{p}{fraction of data to resample.}

\item{sample_weight}{numeric vector with the weights of samples for bootstrap resampling. For classification, if 2 values are given, the 1st one is assumed to be for the positive class (classpos argument).}

\item{in_parallel}{if TRUE, the function is run in parallel}

\item{n_cores}{if in_parallel = TRUE, and no cluster has been passed: number of cores to use}

\item{cluster}{the cluster to use to run the function in parallel}

\item{...}{arguments to be passed to the model2DE funtion.}
}
\value{
A list with the row numbers of partitioned data, the rules originally extracted from the model, a list with results from each bootstrap (use stabilitySelection to obtain the stable decison ensemble).
}
\description{
Wrapper around the model2DE function to run it on several bootstrap resamples.
}
