% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endoR.R
\docType{data}
\name{titanic}
\alias{titanic}
\title{This is data to be included in my package}
\format{
An object of class \code{data.frame} with 2207 rows and 9 columns.
}
\usage{
titanic
}
\description{
This is data to be included in my package
}
\references{
\url{https://cran.r-project.org/package=titanic}
}
\author{
Paul Hendricks \email{paul.hendricks.2013@owu.edu}
}
\keyword{datasets}
